% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adem-package.R
\docType{package}
\name{adem-package}
\alias{adem-package}
\alias{_PACKAGE}
\title{adem: Automated Detection of Excess Mortality}
\description{
This package facilitates automated detection of excess mortality (adem) using an innovative algorithm based on generalized linear models.
}
\author{
\strong{Maintainer}: Kasper Schou Telkamp \email{ksst@ssi.dk} (\href{https://orcid.org/0009-0001-5126-0190}{ORCID})

}
\keyword{internal}
