% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modelling.R
\name{data_modelling}
\alias{data_modelling}
\title{Preprocess Data for Modelling}
\usage{
data_modelling(data, start, end, frequency)
}
\arguments{
\item{data}{A numeric vector or time series data representing the observations.}

\item{start}{A two-element numeric vector specifying the start date in the format c(year, period).}

\item{end}{A two-element numeric vector specifying the end date in the format c(year, period).}

\item{frequency}{The number of observations per year (e.g., 12 for monthly data).}
}
\value{
data frame with formatted data, including columns for 'deaths', 'year', and 'period'.
}
\description{
Preprocess and format time series data for modeling.
}
\examples{
# Load Monthly Deaths from Lung Disease in the UK from MASS
deaths <- MASS::deaths

# Define start and end dates
start_date <- c(1974, 1)
end_date <- c(1979, 12)

# Preprocess the data
formatted_data <- data_modelling(data = deaths, start = start_date, end = end_date, frequency = 12)
}
