% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_append.R
\name{delay_append}
\alias{delay_append}
\title{Title}
\usage{
delay_append(t, opt_par, delay_distribution)
}
\arguments{
\item{t}{A numeric vector of time points at which to calculate the CDF.}

\item{opt_par}{numeric vector of optimized parameters for the chosen delay distribution.
For "geometric" distribution, it should contain a single parameter (success probability).
For "negative.binomial" distribution, it should contain two parameters (size and success probability).}

\item{delay_distribution}{A character string specifying the delay distribution to be used.
It should be either "geometric" or "negative.binomial".}
}
\value{
A numeric vector containing the CDF values at the given time points.
}
\description{
This function calculates the cumulative distribution function (CDF) of registration delays
based on a specified delay distribution model.
}
\examples{
# Example 1: Calculate CDF for the geometric model
cdf_geo <- delay_append(t = c(1, 2, 3), opt_par = 0.3, delay_distribution = "geometric")

# Example 2: Calculate CDF for the negative binomial model
cdf_nb <- delay_append(
  t = c(1, 2, 3),
  opt_par = c(5, 0.2),
  delay_distribution = "negative.binomial"
  )
}
