% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_calculation.R
\name{delay_calculation}
\alias{delay_calculation}
\title{Registration Delay Optimization}
\usage{
delay_calculation(par, x, delay_distribution)
}
\arguments{
\item{par}{A numeric vector of parameters to be optimized.}

\item{x}{A data vector or data frame containing the observed delay data.}

\item{delay_distribution}{A character string specifying the delay distribution to be used.
It should be either "geometric" or "negative.binomial".}
}
\value{
An optimization result object as returned by \code{stats::optim}.
}
\description{
This function performs optimization for registration delay estimation using different delay distribution models.
}
\examples{
observed_delays <- rnbinom(100, 5, prob = 0.5)
# Example 1: Optimize the geometric model
result_geo <- delay_calculation(
  par = c(0.1, 0.5),
  x = observed_delays,
  delay_distribution = "geometric"
  )

# Example 2: Optimize the negative binomial model
result_nb <- delay_calculation(
  par = c(1, 0.2),
  x = observed_delays,
  delay_distribution = "negative.binomial"
  )

}
