% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fitting Generalized Linear Models}
\usage{
fit_model(fit_formula, data, start, fit_distribution)
}
\arguments{
\item{fit_formula}{A formula specifying the model to be fit, e.g., y ~ x1 + x2.}

\item{data}{A data frame containing the variables in the formula.}

\item{start}{A starting point for the optimization algorithm (initial parameter values).}

\item{fit_distribution}{A character string specifying the distribution to be used for fitting the model.
Choose between \code{poisson} or \code{negative.binomial}.}
}
\value{
An object of class \code{glm} or \code{glm.nb} depending on the chosen distribution, representing the fitted model.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Fit a statistical model to data using the specified distribution.
}
\examples{
# Load Monthly Deaths from Lung Disease in the UK from MASS
deaths <- MASS::deaths

# Reformat the data
my_data <- data.frame(
deaths = c(deaths),
date = time(deaths),
t = 1:length(deaths)
)
my_data$year <- trunc(my_data$date)
my_data$month <- (my_data$date - my_data$year) * 12 + 1

# Define a model formula
model_formula <- deaths ~ 1 + t + sin(pi/6*month) + cos(pi/6*month)

# Define the starting values
start <- c(
log(mean(my_data$deaths)),
0, 0, 0
)

# Fit a Negative Binomial model
negbin_model <- fit_model(
fit_formula = model_formula,
data = my_data,
start = start,
fit_distribution = "negative.binomial"
)

# Print the fitted model summary
summary(negbin_model)
}
