% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll_geometric.R
\name{nll_geometric}
\alias{nll_geometric}
\title{Geometric Negative Log Likelihood Function}
\usage{
nll_geometric(theta, data)
}
\arguments{
\item{theta}{A numeric scalar}

\item{data}{A scalar or vector of integer values}
}
\value{
A numeric scalar for the negative log likelihood of the geometric density
given the data and theta parameter.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The negative log likelihood of a geometric density given the data and the theta parameter.
}
\details{
The geometric distribution with \code{prob}\eqn{= p} has density \deqn{p(x)=p(1-p)^x} for \eqn{x=0,1,2,\cdots,0<p\leq 1}.

If an element of \code{x} is not integer, the results of \code{nll_geometric} is zero, with a warning.
}
\examples{
nll_geometric(theta = 0.5, data = rgeom(n = 10, prob = 0.5))
}
