% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll_negative.binomial.R
\name{nll_negative.binomial}
\alias{nll_negative.binomial}
\title{Negative Binomial Negative Log Likelihood Function}
\usage{
nll_negative.binomial(theta, data)
}
\arguments{
\item{theta}{A numeric vector}

\item{data}{A scalar or vector of integer values}
}
\value{
A numeric scalar for the negative log likelihood of the geometric density
given the data and theta parameter.
}
\description{
The negative log likelihood of a negative binomial density given the data and the theta parameter.
}
\details{
The negative binomial distribution with \code{size}\eqn{= n} and \code{prob}\eqn{= p} has density \deqn{p(x)=\frac{\Gamma(x+n)}{\Gamma(n)x!}p^n(1-p)^x} for \eqn{x=0,1,2,\cdots,n>0} and \eqn{0<p\leq 1}.

If an element of \code{x} is not integer, the results of \code{nll_negative.binomial} is zero, with a warning.
}
\examples{
nll_negative.binomial(theta = c(1,0.5), data = rnbinom(10, 1, 0.5))
}
