% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_calculation.R
\name{threshold_calculation}
\alias{threshold_calculation}
\title{Threshold Calculator}
\usage{
threshold_calculation(df, fit, sig_value, threshold_method)
}
\arguments{
\item{df}{A data.frame of new data for which predictions and thresholds are calculated.}

\item{fit}{An object of class \code{glm} or \code{negbin}. This object is used to make predictions and calculate thresholds.}

\item{sig_value}{A numeric value between 0 and 1, controlling the significance level of the interval estimates.}

\item{threshold_method}{A character string defining the method used to calculate thresholds. Choose between \code{prediction} or \code{quantile}.}
}
\value{
A data.frame, \code{df}, with predicted values and upper and lower bounds attached.
}
\description{
Calculate prediction intervals or quantile thresholds for a given data frame using a fitted model.
}
\examples{
# Fit a Poisson model
fit <- glm(y ~ 1, family = "poisson", data = data.frame(y = rpois(10, 5)))
# Calculate a prediction and upper and lower prediction interval
threshold_calculation(
  df = data.frame(y = rpois(1, 5)),
  fit = fit,
  sig_value = 0.05,
  threshold_method = "prediction"
  )
}
