% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adem.R
\name{adem}
\alias{adem}
\title{Automated Detection of Excess Mortality}
\usage{
adem(
  data,
  data_delay = NULL,
  fit_formula,
  k,
  sig_value = 0.05,
  threshold_method = c("prediction", "quantile"),
  theta_start,
  delay_distribution = c("geometric", "negative.binomial"),
  delay_par,
  fit_distribution = c("negative.binomial", "poisson"),
  data_start,
  data_end,
  data_frequency,
  units = c("auto", "secs", "mins", "hours", "days", "weeks")
)
}
\arguments{
\item{data}{A data frame containing the registered time series death data.}

\item{data_delay}{A data frame containing delay information.}

\item{fit_formula}{A formula specifying the model to be fit.}

\item{k}{An integer specifying the window size for modeling.}

\item{sig_value}{A numeric value controlling significance levels.}

\item{threshold_method}{A character string specifying threshold calculation method.
It should be either "prediction" or "quantile".}

\item{theta_start}{A numeric vector for initial parameter values in the model.}

\item{delay_distribution}{A character string specifying the distribution for delay modeling.
It should be either "geometric" or "negative.binomial".}

\item{delay_par}{A numeric scalar or vector for initial parameter values in delay modeling.}

\item{fit_distribution}{A character string specifying the distribution for model fitting.
It should be either "poisson" or "negative.binomial".}

\item{data_start}{A two-element numeric vector specifying the start date in the format c(year, period).}

\item{data_end}{A two-element numeric vector specifying the end date in the format c(year, period).}

\item{data_frequency}{An integer specifying the number of observations per year.}

\item{units}{A character string specifying the time units for delay calculation.
Supported units include "secs", "mins", "hours", "days", "weeks", "months", and "years".}
}
\value{
A \link[tibble:tibble-package]{tibble} containing model results, including predictions, thresholds, and parameter estimates.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Perform Automated Detection of Excess Mortality (ADEM) on time series death data.
}
\examples{
# Load Monthly Deaths from Lung Disease in the UK from MASS
deaths <- MASS::deaths

# Define the model formula
fit_formula = deaths ~ 1 + t + sin(pi/6*period) + cos(pi/6*period)

# Set other parameters
k <- 12*2
sig_value = 0.05
threshold_method = "quantile"
theta_start <- c(log(mean(deaths)), 0, 0, 0)
fit_distribution <- "poisson"
data_start <- c(1974, 1)
data_end <- c(1979, 12)
data_frequency <- 12

# Fit the automated detection of excess mortality function
adem_results <- adem(data = c(deaths),
fit_formula = fit_formula,
k = k,
sig_value = sig_value,
threshold_method = threshold_method,
theta_start = theta_start,
fit_distribution = fit_distribution,
data_start = data_start,
data_end = data_end,
data_frequency = data_frequency)

# Print the results
print(adem_results)
}
