% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_calculation.R
\name{threshold_calculation}
\alias{threshold_calculation}
\title{Title}
\usage{
threshold_calculation(df, fit, sig_value, threshold_method)
}
\arguments{
\item{df}{A data.frame of new data.}

\item{fit}{An object of class \code{glm} or \code{negbin}. Predictions and thresholds are made with this object.}

\item{sig_value}{A numeric value between 0 and 1. Controls the significance level of the interval estimates.}

\item{threshold_method}{A character string that defines the method used to calculate the thresholds. Choose between \code{prediction} or \code{quantile}.}
}
\value{
A data.frame, \code{df}, with predicted values, upper and lower bounds attached.
}
\description{
Title
}
\examples{
# Fit a Poisson model
fit <- glm(y ~ 1, family = "poisson", data = data.frame(y = rpois(10, 5)))
# Calculate a prediction and upper and lower prediction interval
threshold_calculation(
  df = data.frame(y = rpois(1, 5)),
  fit = fit,
  sig_value = 0.05,
  threshold_method = "prediction"
  )
}
